move exception logic into its own file
